+++ src/model/commands/display/SetPalette.java
@@ -11,8 +11,8 @@ public class SetPalette extends DisplayCommand{
 
     @Override
     public double execute () throws Exception {
-        double[] colors = {this.getParams()[1], this.getParams()[2], this.getParams()[3]};
-        getModel().setPaletteColors(this.getParams()[0], colors);
+        int[] colors = {(int) this.getParams()[1], (int) this.getParams()[2], (int) this.getParams()[3]};
+        getModel().setPaletteColors((int) this.getParams()[0], colors);
         return this.getParams()[0];
     }
